+++ RELEASES.md
@@ -1,6 +1,6 @@
 # Release 0.2.0 (2021-09-??)
 - Improved documentation for the decleration macros with traits they implement.
-- Breaking change on declare_array_complex!
+- Breaking change: declare_array_complex now implements real array as well!
 
 ## Complex arrays
 - Added traits for real and imag.